pyproject.toml
[tool.poetry]
name = "letusgo"
version = "0.1.0"
description = ""
authors = ["이한영 <dev@lhy.kr>"]

[tool.poetry.dependencies]
python = "^3.8"

# Django
django = "^3.1"
django-admin-sortable2 = "^0.7.7"
django-aid = "^0.1.5"
django-cleanup = "^4.0"
django-cors-headers = "^3.1"
django-extensions = "^2.2"
django-filter = "^2.2"
django-markdownx = "^3.0.1"
django-modeladmin-reorder = "^0.3.1"
django-phonenumber-field = "^3.0"
django-secrets-manager = "^0.1.13"
django-storages = "^1.7"
easy-thumbnails = "^2.6"
jinja2 = "^2.10"

# DRF
djangorestframework = "^3.10"
djangorestframework-api-key = "^1.4"
djangorestframework-camel-case = "^1.1"
djangorestframework-filters = ">=1.0.0.dev2"
django-rest-auth = "^0.9.5"
drf-writable-nested = "^0.5.1"
drf-yasg = "^1.17"

# Test
codecov = "^2.1.8"
pytest-cov = "^2.7"
pytest-django = "^3.5"
model_bakery = "^1.1.1"

# Error tracking
sentry-sdk = "^0.16.3"

# Others
phonenumberslite = "^8.12.7"
pillow = "^7.2"
psycopg2-binary = "^2.8"
gevent = "^1.4"
gunicorn = "^20.0.4"

[tool.poetry.dev-dependencies]
django-debug-toolbar = "^2.2"
django-sslserver = "^0.22"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
